Add: scale variation
Added in new class to hold additional histogram values for the max
and min of the scale variation envelope. Wrote a simple function
that works to perform the scale variation given a central histo and
a set of other histograms that we vary against. It does not check
that it makes sense to compare the observables yet and it can also
obviously (by design) only handle a single observable at once,
otherwise one need to check observables across all inputs.