Add: change scale variation to accept many inputs
Changed the way scale variation is performed. Now needs switches to
break up sets of arguments. --central_scale are the histograms that
are used as the central scales. --variation_scales are the histos
that are used for the variation, they are matched up by observable
name to the central scales, if they cannot be matched they are not
used. All variations that can be done will be.