Add: Option to compute the efficiency
Added in an option to compute the efficiency, it's just division
though. I wonder if a better approach here might be to write an
interface that allows the user to perform all of the implemented
operations on any given histogram. However efficiency is common
enough in what I am doing to warrant its own option.